* fixed that stupid vs2003 warning
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2286 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/BattlegroundHandler.cpp
@@ -101,7 +101,7 @@ void WorldSession::HandleAreaSpiritHealerQueryOpcode(WorldPacket &recv_data)
 		return;
 	
 	uint32 restime = _player->m_bg->GetLastResurrect() + 30;
-	if(UNIXTIME > restime)
+	if((uint32)UNIXTIME > restime)
 		restime = 1000;
 	else
 		restime = (restime - (uint32)UNIXTIME) * 1000;
